requirements.txt
erlang-py==2.0.1
pylint
urllib3